jimt/CT-822 - new FileRow component
